requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
protobuf==4.24.1
requests==2.31.0
urllib3==2.0.4
websocket-client==1.6.2
PyYAML==6.0.1
